fix: 🐛 profile page with no auth
